\name{mcFastMergeConsensus-package}
\alias{mcFastMergeConsensus-package}
\alias{mcFastMergeConsensus}
\docType{package}
\title{
Multicore fast and merged consensus clustering.
}
\description{
Package contains methods for performing heavy consensus clustering algorithms using multicore processors. One (mcmergecons) implements parallel version of Merge Consensus, originally found in package constClust. Other (mcfastmergecons) implements Fast Merge Consensus Clustering that is a parallel version of sped-up of Merged Consensus Clustering. This package was based on clusterCons package.
}
\details{
\tabular{ll}{
Package: \tab mcFastMergeConsensus\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-06-19\cr
License: \tab GPL-3\cr
}
}
\author{
Tadeusz Tagowski

Maintainer: Tadeusz Tagowski <tadtagov@gmail.com>
}
\references{
Stefano Monti, Pablo Tamayo, Jill Mesirov, and Todd Golub. Consensus clustering: a
resampling-based method for class discovery and visualization of gene expression microar-
ray data. Machine learning, 52(1-2):91–118, 2003.

Raffaele Giancarlo, Filippo Utro, et al. Speeding up the consensus clustering methodology
for microarray data analysis. Algorithms for Molecular Biology, 6(1):1, 2011.

T Ian Simpson, J Douglas Armstrong, and Andrew Jarman. Merged consensus clustering
to assess and improve class discovery with microarray data. BMC bioinformatics, 11(1):590,
2010.

}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
